Bugfix in kms.tf
Without this patch applied, terraform apply would fail with:

  Error: MalformedPolicyDocumentException: The new key policy will not allow you to update the key policy in the future.
        status code: 400, request id: 39345ff4-12d5-48db-9197-21b849dac8be

On further investigation, I found that logic in the locals section was
making assumptions about the caller identity.

In my case, the caller identity was:

  arn:aws:iam::111111111111:user/alex

On further investigation, it appeared there was no need for the logic
around this anyway, thus this patch.